Added Step Cavity Case

+++ data_utils/utils.py
@@ -89,7 +89,7 @@ def parse_arguments():
     parser.add_argument('--dir'         , type=str  , default='test_dir')
     parser.add_argument('--path'        , type=str  , default= r'C:\Users\Noahc\Documents\USYD\PHD\8 - Github\GNOT\data\steady_cavity_case_b200_maxU100ms_simple_normalized.npy')
     parser.add_argument('--epochs'      , type=int  , default=1)
-    parser.add_argument('--sub_x'       , type=int  , default=8)
+    parser.add_argument('--sub_x'       , type=float, default=8)
     parser.add_argument('--inference'   , type=int  , default=1)
     parser.add_argument('--layers'      , type=int  , default=3)
     parser.add_argument('--n_hidden'    , type=int  , default=32)
